Modulo magic
Use the modulo operation to find the barber time in constant space and
linear time

- This works on the premise that on time 0, each barber starts to serve
a customer respectively. Then on each increment a barber can take a new
customer if the current time modulo the barber capacity is 0.
- The next improvement is that we don't need to check this on each
increment of time. We can just use the lowest common factor, if there is
one and jump ahead with those chunks of time.
- The complexity is linear because this operates with `n` steps * `k`
increments of time. `n` is the number of steps till the customer gets
their turn and `k` is a constant increment which is either the lowest
common factor or 1